prueba

+++ src/Archivios.java
@@ -9,7 +9,7 @@ public class Archivios {
         try(BufferedReader bufferedReader = new BufferedReader(new FileReader("C:\\Users\\Usuario\\IdeaProjects\\TpProgramacion\\src\\datos.txt"));){
             while ((line = bufferedReader.readLine()) != null) {
                 String[] ciudadano = line.split(",");
-                ciudadanos.add(new Ciudadano((Integer.parseInt(ciudadano[0])),(Integer.parseInt(ciudadano[1])),(ciudadano[2])));
+                ciudadanos.add(new Ciudadano((Double.parseDouble(ciudadano[0])),(Double.parseDouble(ciudadano[1])),(ciudadano[2])));
             }
         }catch (IOException e){
             System.out.println(e.getMessage());
