+++ src/Inicio.java
@@ -5,8 +5,8 @@ public class Inicio {
             int b = Scanner.getInt("Seleccione el numero de su respuesta: ");
             switch (b) {
                 case 1:
-                    int a = Scanner.getInt("Ingrese su CUIL: ");
-                    int c =Scanner.getInt("Ingrese su Telefono: ");
+                    Double a = Scanner.getDouble("Ingrese su CUIL: ");
+                    Double c =Scanner.getDouble("Ingrese su Telefono: ");
                     buscarUsuario(a, c, ciudadanos, administradores, eventos);
                     System.exit(0);
                     break;
@@ -21,15 +21,15 @@ public class Inicio {
         }
     }
     public static void crearCuenta(Ciudadanos ciudadanos, Evento eventos){
-        Integer cuil = Scanner.getInt("Ingrese su cuil: ");
-        Integer telefono = Scanner.getInt("Ingrese su numero de telefono: ");
+        Double cuil = Scanner.getDouble("Ingrese su cuil: ");
+        Double telefono = Scanner.getDouble("Ingrese su numero de telefono: ");
         String zona = Scanner.getString("Ingrese su zona: ");
         Ciudadano ciudadano = new Ciudadano(cuil, telefono, zona);
         ciudadanos.add(ciudadano);
         Menus.menuCiudadanos(ciudadano, ciudadanos, eventos);
     }
 
-    public static void buscarUsuario(Integer cuil, Integer telefono, Ciudadanos ciudadanos, Administradores administradores, Evento eventos){
+    public static void buscarUsuario(Double cuil, Double telefono, Ciudadanos ciudadanos, Administradores administradores, Evento eventos){
         for (int i = 0;  i < ciudadanos.size(); i++){
             if(cuil.equals(ciudadanos.get(i).getCUIL()) && telefono.equals(ciudadanos.get(i).getTelefono())){
                 Menus.menuCiudadanos(ciudadanos.get(i), ciudadanos, eventos);
